requirements.txt
asgiref==3.7.2
boto3==1.33.2
botocore==1.33.2
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
django-resized==1.0.2
idna==3.6
jmespath==1.0.1
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.8.1
six==1.16.0
sqlparse==0.4.4
sweetify==2.3.1
typing_extensions==4.8.0
urllib3==2.0.7